Implement suggested fixes (complexity simplification, seperate token/lexer files)
